Sync Repository Labels (#7)
# Pull Request

## Description

This pull request includes changes to configure and synchronize GitHub
labels using a new workflow. The most important changes are the addition
of a configuration file for labels and the creation of a GitHub Actions
workflow to sync these labels.

### Configuration of GitHub labels:

*
[`.github/other-configurations/labels.yml`](diffhunk://#diff-c1b32711b0f3da3f5c7007cfaa4d9e94ae5430fea1f4e1256f6210bc6988553dR1-R73):
Added a configuration file defining default GitHub labels, including
labels for bugs, documentation, dependencies, and various code languages
and components.

### Synchronization workflow:

*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6R1-R29):
Created a GitHub Actions workflow to automatically sync labels based on
the configuration file whenever changes are pushed to the main branch.