Add CodeLimit Job (#5)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow in
the `.github/workflows/code-checks.yml` file. The most important change
is the addition of a `run-code-limit` job to run the CodeLimit tool.

New job addition:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R42-R56):
Added a new job named `run-code-limit` to the workflow, which runs the
CodeLimit tool using the `getcodelimit/codelimit-action@v1` action.