Markdown Link Checker (#6)
# Pull Request

## Description

This pull request introduces a new workflow to check for broken Markdown
links in the repository. The most important changes include the addition
of a configuration file for the link checker and the inclusion of a new
job in the GitHub Actions workflow.

Configuration for link checker:

*
[`.github/other-configurations/.linkspector.yml`](diffhunk://#diff-f9691f23ea6c4c34ecbaa23a98721e0aa3c7a94f2fcdd38f7b99fe863c5f6305R1-R7):
Added configuration file specifying directories to check, files to
exclude, and status codes to consider as alive.

GitHub Actions workflow update:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R43-R60):
Added a new job `check-markdown-links` to the workflow to use the
`UmbrellaDocs/action-linkspector` action for checking Markdown links.