Add Dependency Review (#9)
# Pull Request

## Description

This pull request includes changes to the GitHub Actions workflow to add
a new dependency review job. The most important change is the addition
of a new job named `dependency-review` in the
`.github/workflows/pull-request-tasks.yml` file.

Workflow enhancements:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR37-R48):
Added a new job `dependency-review` to perform a dependency review using
the `actions/dependency-review-action@v4.5.0`. This job checks out the
repository and runs a dependency review action.