Add Zizmor Scanning (#10)
# Pull Request

## Description

This pull request includes an update to the GitHub Actions workflow to
add a new job for checking GitHub Actions with Zizmor.

Changes to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R75-R99):
Added a new job `run-zizmor` that checks GitHub Actions with Zizmor,
including steps for checkout, installing the latest version of `uv`,
running Zizmor, and uploading the SARIF file.