Add CodeQL Analysis (#11)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow
for performing CodeQL analysis. The most important change is the
addition of the `run-codeql-analysis` job, which is designed to enhance
security and code quality by analyzing the codebase using CodeQL.

Enhancements to the GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R61-R81):
Added a new job `run-codeql-analysis` to perform CodeQL Analysis,
including steps for checking out the repository, initializing CodeQL,
and performing the analysis.