Add Dependabot Dependency Updates (#13)
# Pull Request

## Description

This pull request includes changes to the `.github/dependabot.yml` file
to configure Dependabot for GitHub Actions updates. The most important
change is the addition of a new configuration to automate dependency
updates.

Dependabot configuration:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R1-R18):
Added a new configuration file to automate updates for GitHub Actions
dependencies on a weekly schedule.