Add Security Policy (#16)
# Pull Request

## Description

This pull request includes an important update to the `docs/SECURITY.md`
file to establish a formal security policy. The update outlines the
procedures for reporting vulnerabilities, supported versions, and how
security updates will be communicated.

Key changes:

*
[`docs/SECURITY.md`](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0R1-R25):
Added a new section titled "Security Policy" that includes instructions
for reporting vulnerabilities, a list of supported versions for security
patches, and information on how security updates will be communicated to
users.