Add Code of Conduct (#17)
# Pull Request

## Description

This pull request includes the addition of a new Code of Conduct for the
project. The document outlines the expectations for behavior within the
community, the responsibilities of community leaders, and the
enforcement guidelines for handling violations.

Key changes include:

* Addition of a comprehensive Code of Conduct document
(`docs/CODE_OF_CONDUCT.md`).

Details of the new Code of Conduct:

* **Our Pledge**: Commitment to a harassment-free experience for
everyone.
* **Our Standards**: Examples of positive and unacceptable behavior.
* **Enforcement Responsibilities**: Community leaders' roles in
maintaining standards.
* **Scope**: Applicability of the Code of Conduct within community
spaces and public representation.
* **Enforcement**: Reporting and handling of unacceptable behavior.
* **Enforcement Guidelines**: Consequences for violations, ranging from
a correction to a permanent ban.