Add README.md (#18)
# Pull Request

## Description

This pull request includes a new `README.md` file for the repository,
which provides information about AWS timing scripts. The most important
changes include the addition of a description, prerequisites,
instructions for running the scripts, and a list of available scripts.

Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R1-R22):
Added a description of the repository, prerequisites, instructions for
running the scripts, and a list of available scripts.