Fix Security Policy Links (#34)
# Pull Request

## Description

This pull request includes updates to the `docs/SECURITY.md` file to
correct the repository links for reporting vulnerabilities and viewing
security advisories.

Changes to security documentation:

*
[`docs/SECURITY.md`](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0L9-R9):
Updated the link for reporting vulnerabilities to point to the correct
repository.
*
[`docs/SECURITY.md`](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0L23-R23):
Updated the link for viewing security advisories to point to the correct
repository.