Add CodeQL configuration (#35)
# Pull Request

## Description

This pull request includes updates to the CodeQL configuration and
workflow files to enhance the security and quality checks. The most
important changes are:

CodeQL configuration updates:

*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beR1-R3):
Added a query filter to exclude the `actions/unpinned-tag` query.

Workflow updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R83):
Updated the CodeQL analysis job to use the new configuration file.