Update Dependency Review (#37)
# Pull Request

## Description

This pull request includes updates to the
`.github/workflows/pull-request-tasks.yml` file to enhance the
dependency review process. The most important changes include adding
permissions for pull requests and configuring the dependency review
action to comment on failures.

Enhancements to dependency review process:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR41-R42):
Added `pull-requests: write` permission to the `dependency-review` job.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR51-R52):
Configured the `dependency-review-action` to comment on the pull request
only when there is a failure.