Check Justfile Format (#39)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow
for checking the format of the `Justfile`. The most important change is
the addition of the `check-justfile-format` job, which ensures that the
`Justfile` format is validated as part of the CI process.

Changes to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R63-R76):
Added a new job `check-justfile-format` to validate the format of the
`Justfile` using the `just format-check` command.