Remove CodeQL Configuration (#58)
# Pull Request

## Description

This pull request updates the CodeQL configuration by removing a custom
query filter and simplifying the workflow configuration. The most
important changes include the removal of an exclusion filter for
unpinned tags and the elimination of a reference to a custom CodeQL
configuration file.

### CodeQL Configuration Updates:

*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beL1-L3):
Removed the exclusion filter for the `actions/unpinned-tag` query, which
was previously used to suppress warnings about unpinned tags in GitHub
Actions.

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L99):
Removed the `config-file` parameter referencing the custom CodeQL
configuration file, simplifying the workflow to use the default CodeQL
settings.