Fix Zizmor SARIF Command (#77)
# Pull Request

## Description

This pull request updates the workflow and task definitions to improve
the integration and execution of `zizmor` checks. The key changes
include replacing the tool used to run `zizmor` commands, introducing a
new setup action for `Just`, and updating dependencies in the workflow
configuration.

### Workflow Updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L130-R137):
Replaced the `uvx` command with `just` for running `zizmor` checks,
added a setup step for `Just`, and updated the `upload-sarif` action to
a newer version (`v3.28.17`).

### Task Definition Updates:
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL47-R51):
Updated the `zizmor-check` task to use `uvx` instead of the older
`zizmor` command and added a new `zizmor-check-sarif` task for
generating SARIF output.