Remove `.editorconfig-checker.json` (#86)
# Pull Request

## Description

This pull request removes the configuration file for the
`.editorconfig-checker` tool and updates the GitHub Actions workflow to
stop referencing it. These changes simplify the project's configuration
and workflow setup.

### Configuration cleanup:

*
[`.github/super-linter-configurations/.editorconfig-checker.json`](diffhunk://#diff-926aaab00380cdd79b8c5069e4eb6d6a7b0a730b2cfd8254e271be7844a0bd10L1-L10):
Removed the `.editorconfig-checker` configuration file, as it is no
longer needed.

### Workflow updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L35):
Removed the `EDITORCONFIG_FILE_NAME` environment variable to reflect the
removal of the `.editorconfig-checker` configuration.